Adding ROM version
- Change FW_REV_ID[0] 15:0 to be the ROM version
- Change FMC version to be 16 bits and shift to FW_REV_ID[0] 31:16
- Adding/updating documentation for FIPS commands

+++ rom/dev/src/flow/cold_reset/fw_processor.rs
@@ -127,7 +127,8 @@ impl FirmwareProcessor {
         report_boot_status(FwProcessorFirmwareDownloadTxComplete.into());
 
         // Update FW version registers
-        env.soc_ifc.set_fmc_fw_rev_id(manifest.fmc.version);
+        // Truncate FMC version to 16 bits (no error for 31:16 != 0)
+        env.soc_ifc.set_fmc_fw_rev_id(manifest.fmc.version as u16);
         env.soc_ifc.set_rt_fw_rev_id(manifest.runtime.version);
 
         // Get the certificate validity info
